refactor: Pokemon IDs are now unique and can't be `0` (#1314)
- `BattleScene#nextPokemonID` added, which is the next ID that will
be assigned to a Pokémon. This is a simple incrementing integer.

- `BattleScene#getNextPokemonID()` added to get the next ID to
assign to a Pokémon and then increment `nextPokemonID`.

- `BattleScene#updateNextPokemonID()` added which sets `nextPokemonID`
to the max of the passed in value
and the current value of `nextPokemonID`.
Used when loading from a save to prevent ID collisions
from pre-existing pokemon.

- The `Pokemon` constructor calls `BattleScene#getNextPokemonID()`
to assign an ID when not loading from a save.

- `Pokemon#personalityValue` added which is a non-unique
random integer from `0` to `2^32 - 1` that is used to
determine shininess instead of the Pokémon's ID.
See https://bulbapedia.bulbagarden.net/wiki/Personality_value#Shininess